# swagger.yaml
openapi: 3.0.3
info:
  title: EPOS Backoffice RESTful APIs
  description: This is the Backoffice RESTful APIs Swagger page.
  contact:
    email: apis@lists.epos-ip.org
  license:
    name: MIT License
    url: https://epos-ci.brgm.fr/epos/WebApi/raw/master/LICENSE
  version: 2.0.0
servers:
  - url: http://localhost:8080/api/backoffice/v1
    description: Inferred Url
paths:
  /user:
    post:
      tags:
        - V3 - User
      summary: "Create User instance"
      description: ""
      operationId: "userpost"
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/user'
      responses:
        '200':
          description: ok.
        '201':
          description: Created.
        '204':
          description: No content.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
    put:
      tags:
        - V3 - User
      summary: "Update Data Product"
      description: ""
      operationId: "userputstate"
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/user'
      responses:
        '200':
          description: ok.
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
  /user/{instance_id}:
    get:
      tags:
        - V3 - User
      summary: "Retrive Data Product"
      description: ""
      operationId: "userget"
      parameters:
        - in: "query"
          name: "userId"
          description: "eduIdentifier"
          required: false
          schema:
            type: string
        - in: "path"
          name: "instance_id"
          description: "Id of the object"
          required: true
          schema:
            type: string
        - in: "query"
          name: "available_section"
          description: "If it is flagged as true, returns the user (or the list of users) populated with the set of section accessible that user"
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userList'
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
    delete:
      tags:
        - V3 - User
      summary: "Delete Data Product"
      description: ""
      operationId: "userdelete"
      parameters:
        - in: "path"
          name: "instance_id"
          description: "instanceId of the object"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ok.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '204':
          description: No content.
        '301':
          description: Moved Permanently.
        '400':
          description: Bad request.
        '401':
          description: Token is missing or invalid
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: 'Invalid input'
        '500':
          description: 'Internal Server Error'
        '501':
          description: 'Not Implemented'
      security:
        - aai: []
...
components:
  schemas:
...
    user:
      required:
        - UID
      type: object
      properties:
        UID:
          type: string
          example: uid
          description: 'This property contains the main identifier for the Data Product, e.g. the URI or other unique identifier in the context of the Catalogue.'
    userList:
      type: array
      items:
        $ref: '#/components/schemas/user'
...
  securitySchemes:
    aai:
      type: apiKey
      in: header
      name: Authorization